Use the UpdateKernel for the updatedb commands.
